Generalizing zsh prompt colors
I've come across a couple of issues with using %F{}/%f to set the colors
in my prompt:

1. The %F{}/%f syntax wasn't supported in the 4.2.x series, which is
still in use on CentOS4 and CentOS5.

2. The %F{}/%f syntax seems to only support 16 colors in the zsh 4.3.10
that I compiled for one server, (while it supports 256 colors in the
same version on another server). It's a pain in the ass to recompile
this again (especially since I don't have access to install *-dev
packages that may be needed).

Therefore, I'm just going to skip over using the %F{}/%f syntax
altogether, and just build a set of variables to easily dealing with the
color / text decoration escape codes.